Overriding python3 default behavior of printing floats, so that integers are printed.
One of the steps of creating input files for aaml is to indicate the sequence length. However, that sequence length cannot contain a decimal. This poses a problem because the default behavior of python3 is to print numbers as floats. This change overrides the default behavior so that an integer is printed instead.

+++ utils.py
@@ -2303,7 +2303,7 @@ def prep_paml_files(orthogroup, indir, outdir, foreground, phylogeny_file, test_
         
             
     if foreground == "aaml_blengths":
-        outfile.write("%s %s\n" % (len(seq_dic), len(rec.seq) / 3))
+        outfile.write("%s %s\n" % (len(seq_dic), int(len(rec.seq) / 3)))
     elif test_type not in ["RELAX", "aBSREL"]:
         outfile.write("%s %s\n" % (len(seq_dic), len(rec.seq)))
     for species, sequence in list(seq_dic.items()):
